+++ pv/data/logic.cpp
@@ -30,6 +30,24 @@ using std::vector;
 namespace pv {
 namespace data {
 
+const QColor Logic::SignalColors[10] =
+{
+	QColor(0x16, 0x19, 0x1A),	// Black
+	QColor(0x8F, 0x52, 0x02),	// Brown
+	QColor(0xCC, 0x00, 0x00),	// Red
+	QColor(0xF5, 0x79, 0x00),	// Orange
+	QColor(0xED, 0xD4, 0x00),	// Yellow
+	QColor(0x73, 0xD2, 0x16),	// Green
+	QColor(0x34, 0x65, 0xA4),	// Blue
+	QColor(0x75, 0x50, 0x7B),	// Violet
+	QColor(0x88, 0x8A, 0x85),	// Grey
+	QColor(0xEE, 0xEE, 0xEC),	// White
+};
+
+const char *Logic::InvalidSignal = "\"%1\" isn't a valid logic signal";
+
+// const SignalBase::ChannelType math_channel_type = SignalBase::LogicMathChannel;
+
 Logic::Logic(unsigned int num_channels) :
 	SignalData(),
 	samplerate_(1),  // Default is 1 Hz to prevent division-by-zero errors
