+++ pv/data/signalbase.hpp
@@ -91,10 +91,11 @@ class SignalBase : public QObject, public enable_shared_from_this<SignalBase>
 
 public:
 	enum ChannelType {
-		AnalogChannel = 1, ///< Analog data
-		LogicChannel,  ///< Logic data
-		DecodeChannel, ///< Protocol Decoder channel using libsigrokdecode
-		MathChannel    ///< Virtual channel generated by math operations
+		AnalogChannel = 1,	///< Analog data
+		LogicChannel,		///< Logic data
+		DecodeChannel,		///< Protocol Decoder channel using libsigrokdecode
+		AnalogMathChannel,	///< Virtual channel generated by analog math operations
+		LogicMathChannel	///< Virtual channel generated by logic math operations
 	};
 
 	enum ConversionType {
@@ -112,9 +113,6 @@ public:
 		NoPreset = -1,     ///< Conversion uses custom values
 		DynamicPreset = 0  ///< Conversion uses calculated values
 	};
-
-	static const QColor AnalogSignalColors[8];
-	static const QColor LogicSignalColors[10];
 
 private:
 	static const int ColorBGAlpha;
@@ -255,6 +253,11 @@ public:
 	 * Get the primary internal data object, i.e. the data that was acquired from the device.
 	 */
 	shared_ptr<pv::data::SignalData> data() const;
+
+	/**
+	 * Get the data as a typed object
+	 */
+	template <typename T> shared_ptr<T> typed_data() const;
 
 	/**
 	 * Determines whether a given segment is complete (i.e. end-of-frame has
@@ -425,6 +428,17 @@ protected:
 
 	QString error_message_;
 };
+
+/**
+ * Get the data as a typed object
+ */
+template <typename T> shared_ptr<T> SignalBase::typed_data() const
+{
+	if (!data_)
+		return nullptr;
+
+	return std::dynamic_pointer_cast<T>(data_);
+}
 
 } // namespace data
 } // namespace pv
